+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog
 ----------------
 
 - Update `catalogView.dtml` to changed behavior of empty searches
-  (`#101 <https://github.com/zopefoundation/Products.ZCatalog/issues/101>`_).
+  (`#102 <https://github.com/zopefoundation/Products.ZCatalog/issues/102>`_).
 
 - Fix case where index value is changed to None after previously being indexed.
   (`#100 <https://github.com/zopefoundation/Products.ZCatalog/issues/100>`_)