+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 5.2 (unreleased)
 ----------------
 
+- Add new method ``searchAll`` to perform a search for all documents.
+
+- Document ``getAllBrains`` and ``searchAll`` in the interface.
+
 - Update `catalogView.dtml` to changed behavior of empty searches
   (`#102 <https://github.com/zopefoundation/Products.ZCatalog/issues/102>`_).